feat: 생성된 롤링페이퍼 페이지의 카드 리스트 컴포넌트 UI 구현
